Added embedded database
As it currently stands:
The program creates a database called Inventory,
Creates 5 tables (Employees, Vehicles, Customers, CurrentRentals, PastRentals),
Adds an entry to the Vehicles table, and prints it to the console just to show you,
Then drops (deletes) all the tables.
The reason for this is that if the tables are not deleted, when you run the program the next time there will be errors because the tables already exist. I'll work on this more later.

Also, this all happens before the gui is even called.
The reason is that we likely won't need to delete all the tables anyway, but they get deleted so that we can run the program again with no problems since the GUI isn't integrated with the DB yet and because it makes the tables every time it runs.